fix: Remove netdata plugin as GPL licence conflicts (#28)
The netdata plugin module go-orchestrator used is GPL licenced. It is also archived but the replacement also appears to be to directly use the go.d.plugin module. However this is also has a GPL license.

To prevent licence conflicts and the GPL licence spreading to any closed source users, this is being removed. It seems fairly unlikely to be in common use. We will jump a minor version number to indicate a breaking change and include on release notes.